backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.8
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
google-generativeai==0.8.3
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
prophet==1.1.6
cmdstanpy==1.2.2
joblib==1.3.2
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.8
# PDF generation library

